# server/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'API Documentation'
  description: 'API documentation for multiple resources'
paths:
  /casts:
    post:
      summary: Create a new cast
      description: Create a new cast entry
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: Cast object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/CastInput'
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
    get:
      summary: Get all casts
      description: Retrieve all cast entries
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /casts/updateCast/{id}:
    put:
      summary: Update a cast by ID
      description: Update an existing cast entry
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the cast to update
          required: true
          type: string
        - name: body
          in: body
          description: Cast object with updated information
          required: true
          schema:
            $ref: '#/definitions/CastInput'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '404':
          description: Not found
  /casts/deleteCast/{id}:
    delete:
      summary: Delete a cast by ID
      description: Delete an existing cast entry
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the cast to delete
          required: true
          type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
...
definitions:
  CastInput:
    type: object
    properties:
      name:
        type: string
      age:
        type: integer
    required:
      - name
      - age
...